Learned about static variables and classes.
If a variable is static, it belongs to the class. For example, if you have a static int in your enemy controller, every instance of an enemy that has the script will have the same value assigned to that static int. Individual instances cannot have different static ints of the same script.